Add json tests for sentry_message
